refactor: move constants to lib/constants for better organization
