+++ src/components/IconItem.svelte
@@ -1,5 +1,5 @@
 <script lang="ts">
-	import { serviceBaseUrl } from "../constants";
+	import { serviceBaseUrl } from "../lib/constants";
 	import type { Icon } from "../models/Icon";
     import * as _iconPreview from '../models/CustomizableIcon';
 	import { customizedIcon } from "../stores/icon-customizations.store";
